fix the build and move types to d.ts
